Rename images/musicbrainz.h to fix build on case-insensitive filesystems
